refactor: move status related functions to reconcile
