+++ shardingsphere-operator/pkg/controllers/compute_node_controller.go
@@ -26,12 +26,12 @@ import (
 	"github.com/apache/shardingsphere-on-cloud/shardingsphere-operator/pkg/kubernetes/configmap"
 	"github.com/apache/shardingsphere-on-cloud/shardingsphere-operator/pkg/kubernetes/deployment"
 	"github.com/apache/shardingsphere-on-cloud/shardingsphere-operator/pkg/kubernetes/service"
+	reconcile "github.com/apache/shardingsphere-on-cloud/shardingsphere-operator/pkg/reconcile/computenode"
 
 	"github.com/go-logr/logr"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
@@ -300,10 +300,6 @@ func (r *ComputeNodeReconciler) reconcileStatus(ctx context.Context, cn *v1alpha
 		return err
 	}
 
-	/*
-		status := reconcileComputeNodeStatus(podlist, service)
-		rt.Status = *status
-	*/
 	reconcileComputeNodeStatus(podlist, service, rt)
 	fmt.Printf("status conditions: %#v\n", rt.Status.Conditions)
 
@@ -343,262 +339,6 @@ func isTrueReadyPod(pod *corev1.Pod) bool {
 	return false
 }
 
-func newConditions(conditions []v1alpha1.ComputeNodeCondition, cond *v1alpha1.ComputeNodeCondition) []v1alpha1.ComputeNodeCondition {
-	if conditions == nil {
-		conditions = []v1alpha1.ComputeNodeCondition{}
-	}
-	if cond.Type == "" {
-		return conditions
-	}
-
-	found := false
-	for idx := range conditions {
-		if conditions[idx].Type != cond.Type {
-			continue
-		}
-		conditions[idx].LastUpdateTime = cond.LastUpdateTime
-		conditions[idx].Status = cond.Status
-		found = true
-		break
-	}
-
-	if !found {
-		conditions = append(conditions, *cond)
-	}
-
-	return conditions
-}
-
-func updateReadyConditions(conditions []v1alpha1.ComputeNodeCondition, cond *v1alpha1.ComputeNodeCondition) []v1alpha1.ComputeNodeCondition {
-	return newConditions(conditions, cond)
-}
-
-func updateNotReadyConditions(conditions []v1alpha1.ComputeNodeCondition, cond *v1alpha1.ComputeNodeCondition) []v1alpha1.ComputeNodeCondition {
-	cur := newConditions(conditions, cond)
-
-	for idx := range cur {
-		if cur[idx].Type == v1alpha1.ComputeNodeConditionReady {
-			cur[idx].LastUpdateTime = metav1.Now()
-			cur[idx].Status = v1alpha1.ConditionStatusFalse
-		}
-	}
-
-	return cur
-}
-
-func newConditionUnknown(reason, message string) v1alpha1.ComputeNodeCondition {
-	return newCondition(v1alpha1.ComputeNodeConditionUnknown, reason, message)
-}
-
-func newConditionPending(reason, message string) v1alpha1.ComputeNodeCondition {
-	return newCondition(v1alpha1.ComputeNodeConditionPending, reason, message)
-}
-
-func newConditionDeployed(reason, message string) v1alpha1.ComputeNodeCondition {
-	return newCondition(v1alpha1.ComputeNodeConditionDeployed, reason, message)
-}
-
-func newConditionStarted(reason, message string) v1alpha1.ComputeNodeCondition {
-	return newCondition(v1alpha1.ComputeNodeConditionStarted, reason, message)
-}
-
-func newConditionInitialized(reason, message string) v1alpha1.ComputeNodeCondition {
-	return newCondition(v1alpha1.ComputeNodeConditionInitialized, reason, message)
-}
-
-func newConditionReady(reason, message string) v1alpha1.ComputeNodeCondition {
-	return newCondition(v1alpha1.ComputeNodeConditionReady, reason, message)
-}
-
-func newConditionFailed(reason, message string) v1alpha1.ComputeNodeCondition {
-	return newCondition(v1alpha1.ComputeNodeConditionFailed, reason, message)
-}
-
-func newCondition(t v1alpha1.ComputeNodeConditionType, reason, message string) v1alpha1.ComputeNodeCondition {
-	return v1alpha1.ComputeNodeCondition{
-		Type:               t,
-		Status:             v1alpha1.ConditionStatusTrue,
-		LastUpdateTime:     metav1.NewTime(time.Now()),
-		LastTransitionTime: metav1.NewTime(time.Now()),
-		Reason:             reason,
-		Message:            message,
-	}
-}
-
-func setConditionUnknown(conditions []v1alpha1.ComputeNodeCondition, reason, message string) {
-	setCondition(conditions, v1alpha1.ComputeNodeConditionUnknown, reason, message, true)
-}
-
-func setConditionPending(conditions []v1alpha1.ComputeNodeCondition, reason, message string) {
-	setCondition(conditions, v1alpha1.ComputeNodeConditionPending, reason, message, false)
-}
-
-func setConditionDeployed(conditions []v1alpha1.ComputeNodeCondition, reason, message string) {
-	setCondition(conditions, v1alpha1.ComputeNodeConditionDeployed, reason, message, false)
-}
-
-func setConditionInitialized(conditions []v1alpha1.ComputeNodeCondition, reason, message string) {
-	setCondition(conditions, v1alpha1.ComputeNodeConditionInitialized, reason, message, false)
-}
-
-func setConditionStarted(conditions []v1alpha1.ComputeNodeCondition, reason, message string) {
-	setCondition(conditions, v1alpha1.ComputeNodeConditionStarted, reason, message, false)
-}
-
-func setConditionReady(conditions []v1alpha1.ComputeNodeCondition, reason, message string) {
-	setCondition(conditions, v1alpha1.ComputeNodeConditionReady, reason, message, false)
-}
-
-func setConditionFailed(conditions []v1alpha1.ComputeNodeCondition, reason, message string) {
-	setCondition(conditions, v1alpha1.ComputeNodeConditionFailed, reason, message, true)
-}
-
-func setCondition(conditions []v1alpha1.ComputeNodeCondition, t v1alpha1.ComputeNodeConditionType, reason, message string, exlusive bool) {
-	cond := v1alpha1.ComputeNodeCondition{
-		Type:               t,
-		Status:             v1alpha1.ConditionStatusTrue,
-		LastUpdateTime:     metav1.NewTime(time.Now()),
-		LastTransitionTime: metav1.NewTime(time.Now()),
-		Reason:             reason,
-		Message:            message,
-	}
-
-	var found bool
-	for i := range conditions {
-		if conditions[i].Type == cond.Type {
-			found = true
-			conditions[i] = cond
-		} else {
-			if cond.Type != v1alpha1.ComputeNodeConditionUnknown {
-
-			}
-			if exlusive {
-				conditions[i].LastUpdateTime = cond.LastUpdateTime
-				conditions[i].Status = v1alpha1.ConditionStatusFalse
-			}
-		}
-	}
-
-	// check current conditions
-	if len(conditions) == 0 || !found {
-		conditions = append(conditions, cond)
-	}
-}
-
-func getConditionFromPods(podlist *corev1.PodList) v1alpha1.ComputeNodeCondition {
-	if len(podlist.Items) == 0 {
-		return newConditionUnknown("PodNotFound", "No pod was found")
-	}
-	var cond v1alpha1.ComputeNodeCondition
-	result := map[v1alpha1.ComputeNodeConditionType]int{}
-	for _, p := range podlist.Items {
-		pc := getPreferedConditionFromPod(p)
-		result[pc.Type]++
-	}
-
-	if result[v1alpha1.ComputeNodeConditionUnknown] == len(podlist.Items) {
-		return newConditionUnknown("PodUnknown", "All pods are unknown")
-	}
-
-	if result[v1alpha1.ComputeNodeConditionReady] > 0 {
-		return newConditionReady("PodReady", "Some pods are ready")
-	}
-
-	if result[v1alpha1.ComputeNodeConditionStarted] > 0 {
-		return newConditionStarted("PodStarted", "Some pods are started")
-	}
-
-	if result[v1alpha1.ComputeNodeConditionInitialized] > 0 {
-		return newConditionInitialized("PodInitialized", "Some pods are initialized")
-	}
-
-	if result[v1alpha1.ComputeNodeConditionDeployed] > 0 {
-		return newConditionDeployed("PodDeployed", "Some pods are deployed")
-	}
-
-	if result[v1alpha1.ComputeNodeConditionPending] > 0 {
-		return newConditionPending("PodPending", "Some pods are pending")
-	}
-
-	if result[v1alpha1.ComputeNodeConditionFailed] > 0 {
-		return newConditionFailed("PodFailed", "Some pods are failed")
-	}
-
-	return cond
-}
-
-func getPreferedConditionFromPod(pod corev1.Pod) v1alpha1.ComputeNodeCondition {
-	if pod.Status.Phase == corev1.PodUnknown {
-		return v1alpha1.ComputeNodeCondition{
-			Type: v1alpha1.ComputeNodeConditionUnknown,
-		}
-	}
-
-	if pod.Status.Phase == corev1.PodPending {
-		return v1alpha1.ComputeNodeCondition{
-			Type: v1alpha1.ComputeNodeConditionPending,
-		}
-	}
-
-	if pod.Status.Phase == corev1.PodFailed {
-		return v1alpha1.ComputeNodeCondition{
-			Type: v1alpha1.ComputeNodeConditionFailed,
-		}
-	}
-
-	return getPreferedConditionFromPodConditions(pod.Status.Conditions)
-}
-
-func getPreferedConditionFromPodConditions(conditions []corev1.PodCondition) v1alpha1.ComputeNodeCondition {
-	var (
-		sched       bool
-		initialized bool
-		conReady    bool
-		ready       bool
-	)
-
-	for _, c := range conditions {
-		if c.Type == corev1.PodScheduled && c.Status == corev1.ConditionTrue {
-			sched = true
-		}
-		if c.Type == corev1.PodInitialized && c.Status == corev1.ConditionTrue {
-			initialized = true
-		}
-		if c.Type == corev1.ContainersReady && c.Status == corev1.ConditionTrue {
-			conReady = true
-		}
-		if c.Type == corev1.PodReady && c.Status == corev1.ConditionTrue {
-			ready = true
-		}
-	}
-
-	if ready {
-		return v1alpha1.ComputeNodeCondition{
-			Type: v1alpha1.ComputeNodeConditionReady,
-		}
-	}
-
-	if conReady {
-		return v1alpha1.ComputeNodeCondition{
-			Type: v1alpha1.ComputeNodeConditionStarted,
-		}
-	}
-
-	if initialized {
-		return v1alpha1.ComputeNodeCondition{
-			Type: v1alpha1.ComputeNodeConditionInitialized,
-		}
-	}
-
-	if sched {
-		return v1alpha1.ComputeNodeCondition{
-			Type: v1alpha1.ComputeNodeConditionDeployed,
-		}
-	}
-
-	return v1alpha1.ComputeNodeCondition{}
-}
-
 func updateComputeNodeStatusCondition(conditions []v1alpha1.ComputeNodeCondition, cond v1alpha1.ComputeNodeCondition) []v1alpha1.ComputeNodeCondition {
 	var found bool
 	for i := range conditions {
@@ -627,7 +367,7 @@ func updateComputeNodeStatusCondition(conditions []v1alpha1.ComputeNodeCondition
 }
 
 func reconcileComputeNodeStatus(podlist *corev1.PodList, svc *corev1.Service, cn *v1alpha1.ComputeNode) {
-	cond := getConditionFromPods(podlist)
+	cond := reconcile.GetConditionFromPods(podlist)
 	cn.Status.Conditions = updateComputeNodeStatusCondition(cn.Status.Conditions, cond)
 
 	ready := getReadyProxyInstances(podlist)
